requirements.txt
discord.py
python-dotenv
qrcode
pillow
mercadopago